Update patch set 2
Patch Set 2: Code-Review-2

I am not convinced with this because we have to launch different containers to run different test cases. This is not an applicable method. Though it is not good to keep restapi container holding all tests, at least we could keep it for E until an alternative is proposed.

Patch-set: 2
Reviewer: Gerrit User 5441 <5441@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0