Update patch set 2
Patch Set 2:

I do not oppose to remove restapi, but not now.
Though docker slicing is quite good, and the process of running each tier one by one in Jenkins looks reasonable, the issue is for a new user of functest,  he/she should not deal with the process of creating which container according to each test. Instead, all the user needs is to run any test case with a unified entrance. So all spilt containers requires a unified management.
I would suggest the unified entrance be realized first, then we could remove restapi container.

Patch-set: 2