Update patch set 2
Patch Set 2:

As PTL, I take the responsability to go forward and to remove the container. It's already well rated and the former Dockerfile could be reused externally for any purpose. Regarding OPNFV, the container doesn't meet the Docker slicing approach we are promoting and it's simply unused.

Linda, I'm very sorry to remove your -2 to progress. I can't badly decrease your vote to -1 to keep your feedbacks.

Patch-set: 2