Rewrite TempestResourcesManager
It now leverages on OpenStack Shade to allow gating OpenStack master
as asked by TSC. Switching to OpenStack master (OpenStack SDK Rocky)
requires to change only one line.

It stops dedicating the network to the user possibly created because
the support is incomplete (the users have to be registed in
tempest.conf and the multiprocessing has to be enforced).

It also writes the second image and the second flavor in tempest.conf.

Conflicts:
    functest/opnfv_tests/openstack/tempest/tempest.py
    functest/tests/unit/openstack/tempest/test_conf_utils.py
    functest/tests/unit/openstack/tempest/test_tempest.py

Change-Id: I9fe2c90a6073ebc240e9aa8fdf2ef93d5dd5f15f
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>
(cherry picked from commit 9fb702111ae63824a2e293d327437d0486a565d2)