Update patch set 1
Patch Set 1:

Thx Cedric for your quick comments.
Juju bootstrap is supposed to have a default timeout of 10 minutes; which we have not overridden. However, if you have observed the bootstrap command hanging, then it might point to some bug in the command - so we should wrap it in a timeout for safety.

Patch-set: 1
Label: Verified=0