SNAPS via functest with flavor metadata configurable
The required data for the SNAPS flavors are sent through
the config_functest yaml file in order for the SNAPS scenarios to be
configured accordingly.
If the config_functest yaml file includes empty flavor_extra_specs
or there is not included this parameter, the SNAPS scenarios will be
executed using the hardcoded value.

Change-Id: I4ef46d1fe93497d658414170a72c4520d8f37808
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>