[WIP] Config flavor metadata via functest stable/fraser
The required data for the SNAPS flavors are sent through
the config_functest yaml file in order for the SNAPS
scenarios to be configured accordingly.
If the config_functest yaml file includes empty
flavor_extra_specs or there is not included this parameter,
the SNAPS scenarios will be executed using the hardcoded value.

Functest documents are also updated.

Depends-On: https://gerrit.opnfv.org/gerrit/58491

Change-Id: I438b72a01ceb24d25e1f0a8ad4dd3147577e41aa
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>