Update patch set 1
Patch Set 1:

Actually it's a triple nested virtualization. So the cirros VM is running in a virtualized compute node which is inside a big big virtual machine.

I know 60 loops are a lot, but the timeout is still 10 seconds so as soon as the network is available we will connect to the VM without having to wait for all 10 minutes to pass so it should have no visible impact to the existing behavior.

Patch-set: 1
Label: Verified=0