Make synchronous Shade calls
It makes difficult to find the issues (floating ip, boot, ssh
connection, etc...) for a small benefit regarding the current
testcase implementations.

It increases the default floating ip timeout because it has been
reached when running Functest multiple times in parallel. It may
be the root cause when running singlevm vs XCI (in that case,
neutron api and agents are hosted in a nested vm).

Change-Id: Ia97e04902321644c207c59987283754318aee5e5
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>