Conform with OpenStack gates
It adds testcases and updates regexes to conform with OpenStack gates:
  - tempest-full
  - neutron-tempest-plugin-api

It should be noted that tempest-smoke is no longer run in gates and
two testcases are available for tempest-full as we can't manage
Tempest combine.

Criteria is set to 100% for tempest_full (it works in OpenStack
gates).

JIRA: FUNCTEST-1000

Change-Id: I77a0f533255483bdf61f0a5f60c9c437ce843591
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>