Publish Functest daily and gate jobs
As opposed to the Releng ones, they are designed to be easily
executed by endusers as well. They can be considered as a first step
towards a full Functest CI/CD toolchain deployed via docker-compose.

They hugely leverage on templates to load the right yamls per branch
and multijobs to run testcases in parallel. Many steps are proposed
allowing running part of the chain.

Executors have to be increased to allow running them (it was tested
vs 10 and more).

At first, all docker images are downloaded to avoid different
versions between testcases.

They will be improved in a second change to ensure that only one node
is selected and to take all Functest env vars into account.

Change-Id: I99c7b8ee5d0ddbb00102b3147739b5e8b5c65c23
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>