Add concurrency parameter to refstack_defcore tests
In fraser release we observed that refstack tests are executed in parallel
and that might causes problems in CI because without concurrency
parameter it will try to use as many cpus as possible and that might
lead to race condition issues.

This patch is created for backporting purposes to fraser branch
where the problem occurs.

JIRA: DOVETAIL-662

Change-Id: I6cf0ad3a32e88f04abd00d69523aa8ac47c39470
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>