Create script to download images for functest-smoke
The script download_images.sh downloads all the images that are used for
various tiers. For smoke tier, only 3 of them are needed so having this
new script will cut the time to run CI. Also, it gives control over which
images are needed for which tiers to Functest so users of this script
can get the right/needed images as the list updated.

Change-Id: If5909ed12b8293d40ebbadd7be791619e4ffa9cb
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>