Create script to download images for functest-healthcheck
The script download_images.sh downloads all the images that are used for
various tiers. For healthcheck tier, only cirros is needed so having this
new script will cut the time to run CI. Also, it gives control over which
images are needed for which tiers to Functest so users of this script
can get the right/needed images as the list updated.

Change-Id: Ic82ea0988bdecc321f80a0276a3ea3cf96efb49a
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>