Update patch set 6
Patch Set 6: Code-Review-1

This patch work at master as it is. I am -1 this because it needs to be cherry picked to stable/gambia branch and I need to check that libvirt-dev apk package is compatible with libvirt-python pip package.

The problem there is that openstack-queens upper constraint is libvirt-python==3.10 which is incompatible with apk libvirt-dev==4.4, which is present to v3.8 apk alpine repo.

I am trying to find ways to fetch libvirt-dev from alpine apk repo v3.7 which is compatible with libvirt-python constraint of queens.

Patch-set: 6
Label: Code-Review=-1