Deploy Functest CI toolchains
It eases verifying OpenStack by deploying the full testing toolchain
via Ansible. As opposed to the Releng model, all tests can be executed
in parallel.

It leverages on Xtesting Ansible role which currently deploys
Jenkins, Minio and S3www and loads all jobs in Jenkins.
Last components (MongoDB, TestAPI and dashboard) will be deployed in
the next update of the ansible role.

To deploy the testing toolchain:
$ ansible-galaxy install collivier.xtesting
$ ansible-playbook ansible/site.yml

Change-Id: I260806c1d19562891d2ed9c52ba37a82bf60792e
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>