Update patch set 3
Patch Set 3:

> Cristina: An email has been sent. Be free to reply to it.
 > 
 > Functest teams agreed on that change during last Functest meeting
 > and this is needed to achieve our targets.

The email has been sent to the installer projects and functest members, which received no opposition. I too agree with this change. What I am referring to is the fact that the change has not been approved by the TSC (which is the one who sets the release criteria). The merge should be made after the formal approval is received.

Patch-set: 3