Update and enable Cloudify-based testcases
It switches from the Cloudify virtual machine to the community
container which allows deploying vyos_vrouter and cloudify_ims.

Last stable Cloudify OpenStack plugin (2.4.17) is now deployed.
It also updates the Cloudify python package to latest release.

It adds a delay before deleting the blueprint which may be
improved in a second change (it allows enabling the testcase in
the stable release).

Change-Id: Ie48a2c9d2badab0d7ba0b9e7845f453fad335ef7
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>