Update patch set 1
Patch Set 1:

(1 comment)

When a certain testcase fails it is flagged as failed. If there are still cases on the execution list, they won't be executed (nor won't be part of the report) but the execution of next scenario group is started. Good to know when interpreting success/failure figures.

-1 was just for the name/title issue.

Patch-set: 1