Update patch set 2
Patch Set 2: Code-Review-1

> (1 comment)
 > 
 > When a certain testcase fails it is flagged as failed. If there are
 > still cases on the execution list, they won't be executed (nor
 > won't be part of the report) but the execution of next scenario
 > group is started. Good to know when interpreting success/failure
 > figures.
 > 
 > -1 was just for the name/title issue.

I applied changes to name/title and commit message and also rebased to include Cedric's patch.

Speaking about aborting tests we could add logic to compare the test cases which run with the test cases inside yml files. If a test case name is not present inside rally json file then we can assume it was aborted. In the end we can return a list of aborted test cases per category.

What is your opinion on the matter?

Patch-set: 2