Bug on tempest results' parser about skipped TCs
While the test case neutron-tempest-plugin-api is run through dovetail
container, has been noticed that the skipped sub-testcases are not
collected properly.

Those with the format 'skip [xxx]:' are not
collected in the functest_result.txt file.

Therefore, those sub-testcases recognized as Failed by dovetail
container.

The condition has been updated in order to match both format of
skipped testcases ('skip:' and 'skip [xxx]:')

The bug has been noticed in the master branch and backported
to the gambia release.

Change-Id: I11c2ecce0f688e6f1d6f8d0843b8879fd3c75e29
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>