Modify the upstream Clearwater Heat files
It forces a single network for both management and signaling traffic.
It meets the OPNFV deployment requirements and the Functest SUT ones.

The security group will be improved in a second step to first quickly
fix the testcase in gambia. Private IPs may be registered in DNS to
avoid network address translations.

Change-Id: Ic19cf336ac5c2d07c52c6dd37b06271790145cf9
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>