Set cinder service type in rally tests
As a precondition for executing scenarios requiring volume, Rally
expects the presence of service named as "cinder". In a case when
there is only cinderv2 and cinderv3 configured in the system, Rally
fails to validate a testcase even the volume service is working ok.

Testcases in question changed to use v3.

Change-Id: I42e7f846e80d3647df274b2e03add94c9f0a34d6
Signed-off-by: Juha Kosonen <juha.kosonen@nokia.com>
(cherry picked from commit ebc2b41505299846faeb2163d34d820a8ef2bdb2)