Clean orphan security groups
It removes all security groups which are not owned by a project.
This approach seems suitable as it works even if dynamic accounts
(tempest and rally). All other OpenStack resources are still
removed by the underlying framework.

It lists all orphan security groups in use as debug to avoid
misunderstanding the testcase results (it could happen if cloud admin
removes accounts without cleaning the virtual machines).

Change-Id: I033d96e077b4277d59233a3de4ca6d41ed91f743
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>