Add oauth1 as optional Keystone features
It was not listed in Keystone which defacto sets it as optional.
It may be considered as mandatory in a second step.

It should be noted that it's not working in CNTT RI and there is no
true plan to update CNTT RI from the time writing.

Close cntt-n/CNTT#1253

Change-Id: If7ee4efeafc05638173a51ce997ecbdcaabfc79c
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>
(cherry picked from commit 13b8fd9d9f54fdac817b99f4506a1171ab0a9481)