Add tempest_heat (heat-tempest-plugin)
It asks for the Fedora-Cloud-Base image and a second user account.
It should be noted that heat-tempest-plugin duplicates all configs from
tempest.

Minimal image is still Cirros and the first account is the admin one.
Then it conforms with the TempestCommon logic.

AodhAlarmTest is skipped by default because pre conditions are missing.
SoftwareConfigIntegrationTest is skipped as in gates because it
requires a custom image [1]

[1] https://github.com/openstack/heat/blob/master/devstack/lib/heat#L444

Change-Id: I8b39dc65ef3714411cd828e17b95e124f2e90f34
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>