Set concurrency and containers_per_tenant to 1
Most of the rally_full runs failed vs our new SUTs (radosgw) .
This two params fail the testing when higher than 1 (verified
independently).

As the new test config works when hugely tested in //, it makes sens to
merge the new config to fix the gates.

This patch will be safely reverted if the issue is confirmed on SUT.

Change-Id: I222652d484a5e06190c12d45d6c1882de5ab9eec
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>