Update patch set 1
Patch Set 1:

All CNTT specific tests are quickly described in RC (it should be enough for a tempest or rally user).
https://github.com/cntt-n/CNTT/blob/master/doc/ref_cert/lfn/chapters/chapter03.md#3.3.4

Our Functest developper guide is quite huge and possibly ask for a refresh.
Special thanks to the PTLs and the Nokia devs who have contributed to this huge work. 
https://functest.readthedocs.io/en/latest/testing/developer/devguide/index.html

Patch-set: 1