Clone repositories instead of wgetting constraints
It stops leveraging on wget to get constraints as it fails since
opendev was created. Then functest and requirements are no longer
removed for functest-core which simplifies all child Dockerfile.

It also fix the current incompatibilities in OpenStack constraints
ERROR: cliff 3.2.0 has requirement cmd2!=0.8.3,<0.9.0,>=0.8.0, but you'll have cmd2 1.1.0 which is incompatible.
https://github.com/openstack/requirements/blob/master/upper-constraints.txt#L230
https://github.com/openstack/requirements/blob/master/upper-constraints.txt#L416

Change-Id: I7c044208ef26ad047fb841710b422acb6c9f0320
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>